Verify user additional info in a separate component
