+++ apps/api-portal/components/require-additional-info/require-additional-info.spec.tsx
@@ -1,0 +1,9 @@
+import { render } from '@testing-library/react'
+import RequireAdditionalInfo from './require-additional-info'
+
+describe('RequireAdditionalInfo', () => {
+  it('should render successfully', () => {
+    const { baseElement } = render(<RequireAdditionalInfo />)
+    expect(baseElement).toBeTruthy()
+  })
+})
