+++ apps/api-portal/pages/index.tsx
@@ -1,11 +1,9 @@
 import type { NextPage } from 'next'
 // import Layout from 'components/layout'
 import dynamic from 'next/dynamic'
-import { Fragment, useEffect } from 'react'
-import { useRouter } from 'next/router'
 import AccessTokenList from 'components/access-token/access-token-list/access-token-list'
 import AccessTokenCreate from 'components/access-token/access-token-create/access-token-create'
-import { useGetUserApplications } from 'features/user-applications/user-applications.hooks'
+import RequireAdditionalInfo from 'components/require-additional-info/require-additional-info'
 import styles from '../styles/index.module.css'
 
 const Layout = dynamic(() => import('components/layout'), {
@@ -13,17 +11,6 @@ const Layout = dynamic(() => import('components/layout'), {
 })
 
 const Home: NextPage = (context) => {
-  const { isUserApplicationsRequiredInfoCompleted } = useGetUserApplications()
-  const router = useRouter()
-
-  useEffect(() => {
-    if (
-      isUserApplicationsRequiredInfoCompleted !== null &&
-      !isUserApplicationsRequiredInfoCompleted
-    ) {
-      router.push('/signup')
-    }
-  }, [isUserApplicationsRequiredInfoCompleted, router])
   return (
     <Layout>
       <p className={styles.description}>
@@ -31,13 +18,10 @@ const Home: NextPage = (context) => {
         4wings activity tiles. Read more about API access tokens in{' '}
         <a href="https://slatedocs.github.io/slate/#introduction">our documentation</a>
       </p>
-      {isUserApplicationsRequiredInfoCompleted && (
-        <Fragment>
-          <AccessTokenList></AccessTokenList>
-          <AccessTokenCreate></AccessTokenCreate>
-        </Fragment>
-      )}
-      {!isUserApplicationsRequiredInfoCompleted && <div>provide more details</div>}
+      <RequireAdditionalInfo>
+        <AccessTokenList></AccessTokenList>
+        <AccessTokenCreate></AccessTokenCreate>
+      </RequireAdditionalInfo>
     </Layout>
   )
 }
